confidential-ledger-demo/server/requirements.txt
starlette
databases
sqlalchemy
aiofiles
jinja2
#python-multipart
python-dotenv
# Required for blob storage
aiohttp
# TODO: Fix aiohttp speedups
#aiohttp[speedups]
# Required for deploying to Azure
uvicorn
uvloop
httptools
azure-confidentialledger
azure-mgmt-confidentialledger
azure-identity